pyproject.toml
[project]
name = "pdf-chunker"
version = "0.1.0"
description = "Convert PDFs to AI-optimized Markdown chunks with rich metadata"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pymupdf>=1.25.0",
    "pdfplumber>=0.11.0",
    "pydantic>=2.0",
    "click>=8.1",
    "tiktoken>=0.8.0",
]

[project.scripts]
pdf-chunker = "pdf_chunker.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.hatch.build.targets.wheel]
packages = ["src/pdf_chunker"]

[dependency-groups]
dev = [
    "fpdf2>=2.8.7",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "pyyaml>=6.0.3",
]
